+++ src/platform.ts
@@ -1,22 +1,34 @@
 import {API, DynamicPlatformPlugin, Logger, PlatformAccessory, PlatformConfig, Service, Characteristic} from 'homebridge';
 
 import {PLATFORM_NAME, PLUGIN_NAME} from './settings';
-import {DaikinCloudAirConditioningAccessory} from './accessory';
+import {DaikinClimateControlEmbeddedId, daikinAirConditioningAccessory} from './daikinAirConditioningAccessory';
 
-import DaikinCloudController from 'daikin-controller-cloud';
-import path from 'path';
-import fs from 'fs';
+import {DaikinCloudController} from 'daikin-controller-cloud/dist/index.js';
 
-import type * as Device from './../node_modules/daikin-controller-cloud/lib/device.js';
-import type * as DaikinCloud from './../node_modules/daikin-controller-cloud/index.js';
+import {daikinAlthermaAccessory} from './daikinAlthermaAccessory';
+import {resolve} from 'node:path';
+import {DaikinCloudDevice} from 'daikin-controller-cloud/dist/device';
+import {StringUtils} from './utils/strings';
+
+const ONE_SECOND = 1000;
+const ONE_MINUTE = ONE_SECOND * 60;
+
+export type DaikinCloudAccessoryContext = {
+    device: DaikinCloudDevice;
+};
 
 export class DaikinCloudPlatform implements DynamicPlatformPlugin {
-    public readonly Service: typeof Service = this.api.hap.Service;
-    public readonly Characteristic: typeof Characteristic = this.api.hap.Characteristic;
+    public readonly Service: typeof Service;
+    public readonly Characteristic: typeof Characteristic;
 
-    public readonly accessories: PlatformAccessory[] = [];
+    public readonly accessories: PlatformAccessory<DaikinCloudAccessoryContext>[] = [];
 
     public readonly storagePath: string = '';
+    public controller: DaikinCloudController;
+
+    public readonly updateIntervalDelay = ONE_MINUTE * 15;
+    public updateInterval: NodeJS.Timeout | undefined;
+    public forceUpdateTimeout: NodeJS.Timeout | undefined;
 
     constructor(
         public readonly log: Logger,
@@ -25,26 +37,57 @@ export class DaikinCloudPlatform implements DynamicPlatformPlugin {
     ) {
         this.log.debug('Finished initializing platform:', this.config.name);
 
+        this.Service = this.api.hap.Service;
+        this.Characteristic = this.api.hap.Characteristic;
         this.storagePath = api.user.storagePath();
+        this.updateIntervalDelay = ONE_MINUTE * (this.config.updateIntervalInMinutes || 15);
+        this.controller = new DaikinCloudController({
+            oidcClientId: this.config.clientId,
+            oidcClientSecret: this.config.clientSecret,
+            oidcCallbackServerBindAddr: this.config.oidcCallbackServerBindAddr,
+            oidcCallbackServerExternalAddress: this.config.callbackServerExternalAddress,
+            oidcCallbackServerPort: this.config.callbackServerPort,
+            oidcTokenSetFilePath: resolve(this.storagePath, '.daikin-controller-cloud-tokenset'),
+            oidcAuthorizationTimeoutS: 60 * 5,
+        });
 
-        this.api.on('didFinishLaunching', () => {
-            log.debug('Executed didFinishLaunching callback');
-            this.discoverDevices(this.config.username, this.config.password);
+
+        this.api.on('didFinishLaunching', async () => {
+            this.controller.on('authorization_request', (url) => {
+                this.log.warn(`
+                    Please navigate to ${url} to start the authorisation flow. If it is the first time you open this url you will need to accept a security warning.
+                    
+                    Important: Make sure your Daikin app Redirect URI is set to ${url} in the Daikin Developer Portal.
+                `);
+            });
+
+            this.controller.on('rate_limit_status', (rateLimitStatus) => {
+                if (rateLimitStatus.remainingDay && rateLimitStatus.remainingDay <= 20) {
+                    this.log.warn(`[Rate limit remaining calls] Rate limit almost reached, you only have ${rateLimitStatus.remainingDay} calls left today`);
+                }
+                this.log.debug(`[Rate limit remaining calls] today: ${rateLimitStatus.remainingDay}/${rateLimitStatus.limitDay} -- this minute: ${rateLimitStatus.remainingMinute}/${rateLimitStatus.limitMinute}`);
+            });
+
+            await this.discoverDevices(this.controller);
+            this.startUpdateDevicesInterval();
+
         });
     }
 
-    configureAccessory(accessory: PlatformAccessory) {
+    public configureAccessory(accessory: PlatformAccessory<DaikinCloudAccessoryContext>) {
         this.log.info('Loading accessory from cache:', accessory.displayName);
         this.accessories.push(accessory);
     }
 
-    async discoverDevices(username: string, password: string) {
-        let devices: Device[] = [];
+    private async discoverDevices(controller: DaikinCloudController) {
+        let devices: DaikinCloudDevice[] = [];
 
-        this.log.info('--- Daikin info for debugging reasons (enable Debug Mode for more logs v1.6.1) ---');
+        this.log.info('--- Daikin info for debugging reasons (enable Debug Mode for more logs) ---');
+
+        this.log.debug('[Config] User config', this.getPrivacyFriendlyConfig(this.config));
 
         try {
-            devices = await this.getCloudDevices(username, password);
+            devices = await controller.getCloudDevices();
         } catch (error) {
             if (error instanceof Error) {
                 error.message = `Failed to get cloud devices from Daikin Cloud: ${error.message}`;
@@ -55,13 +98,9 @@ export class DaikinCloudPlatform implements DynamicPlatformPlugin {
         devices.forEach(device => {
             try {
                 const uuid = this.api.hap.uuid.generate(device.getId());
-
-                this.log.info('Device found with id: ' + uuid);
-                this.log.info('    name: ' + device.getData('climateControl', 'name').value);
-                this.log.info('    last updated: ' + device.getLastUpdated());
-                this.log.info('    modelInfo: ' + device.getData('gateway', 'modelInfo').value);
-                this.log.info('    config.showExtraFeatures: ' + this.config.showExtraFeatures);
-                this.log.info('    config.excludedDevicesByDeviceId: ' + this.config.excludedDevicesByDeviceId);
+                const climateControlEmbeddedId: DaikinClimateControlEmbeddedId = device.getDescription().deviceModel === 'Altherma' ? 'climateControlMainZone' : 'climateControl';
+                const name: string = device.getData(climateControlEmbeddedId, 'name', undefined).value;
+                const deviceModel: string = device.getDescription().deviceModel;
 
                 const existingAccessory = this.accessories.find(accessory => accessory.UUID === uuid);
 
@@ -77,15 +116,30 @@ export class DaikinCloudPlatform implements DynamicPlatformPlugin {
                     this.log.info('Restoring existing accessory from cache:', existingAccessory.displayName);
                     existingAccessory.context.device = device;
                     this.api.updatePlatformAccessories([existingAccessory]);
-                    new DaikinCloudAirConditioningAccessory(this, existingAccessory);
+
+                    if (deviceModel === 'Altherma') {
+                        new daikinAlthermaAccessory(this, existingAccessory);
+                    } else {
+                        new daikinAirConditioningAccessory(this, existingAccessory);
+                    }
+
                 } else {
-                    this.log.info('Adding new accessory:', device.getData('climateControl', 'name').value);
-                    const accessory = new this.api.platformAccessory(device.getData('climateControl', 'name').value, uuid);
+                    this.log.info('Adding new accessory:', name);
+                    const accessory = new this.api.platformAccessory<DaikinCloudAccessoryContext>(name, uuid);
                     accessory.context.device = device;
-                    new DaikinCloudAirConditioningAccessory(this, accessory);
+
+                    if (deviceModel === 'Altherma') {
+                        new daikinAlthermaAccessory(this, accessory);
+                    } else {
+                        new daikinAirConditioningAccessory(this, accessory);
+                    }
+
                     this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
                 }
             } catch (error) {
+                // eslint-disable-next-line no-console
+                console.error(error);
+
                 if (error instanceof Error) {
                     this.log.error(`Failed to create HeaterCooler accessory from device, only HeaterCooler is supported at the moment: ${error.message}, device JSON: ${JSON.stringify(device)}`);
                 }
@@ -95,69 +149,40 @@ export class DaikinCloudPlatform implements DynamicPlatformPlugin {
         this.log.info('--------------- End Daikin info for debugging reasons --------------------');
     }
 
-    async getCloudDevices(username: string, password: string): Promise<Device[]> {
-        const daikinCloud = await this.initiateDaikinCloudController(username, password);
-        const devices: Device[] = await daikinCloud.getCloudDevices();
-
-        this.log.info(`Found ${devices.length} devices in your Daikin Cloud`);
-
-        if (devices.length === 0) {
-            return devices;
-        }
-
-        const cloudDetails = await daikinCloud.getCloudDeviceDetails();
-        this.log.debug(JSON.stringify(cloudDetails));
-
-        return devices;
+    private async updateDevices() {
+        this.log.debug('Update devices data');
+        await this.controller.updateAllDeviceData();
     }
 
-    async initiateDaikinCloudController(username, password) {
-        let tokenSet;
-        const options = {
-            // eslint-disable-next-line no-console
-            logger: console.log,            // optional, logger function used to log details depending on loglevel
-            logLevel: 'info',               // optional, Loglevel of Library, default 'warn' (logs nothing by default)
-            proxyOwnIp: '192.168.xxx.xxx',  // required, if proxy needed: provide own IP or hostname to later access the proxy
-            proxyPort: 8888,                // required: use this port for the proxy and point your client device to this port
-            proxyWebPort: 8889,             // required: use this port for the proxy web interface to get the certificate and start Link for login
-            proxyListenBind: '0.0.0.0',     // optional: set this to bind the proxy to a special IP, default is '0.0.0.0'
-            proxyDataDir: this.storagePath, // Directory to store certificates and other proxy relevant data to
-            communicationTimeout: 10000,    // Amount of ms to wait for request and responses before timeout
-            communicationRetries: 3,        // Amount of retries when connection timed out
-        };
+    forceUpdateDevices(delay: number = ONE_SECOND * 60) {
+        this.log.debug(`Force update devices data (delay: ${delay}, update pending: ${this.forceUpdateTimeout})`);
 
-        const tokenFile = path.join(this.storagePath, 'daikincloudtokenset.json');
-        this.log.debug(`Write/read Daikin Cloud tokenset from ${tokenFile}`);
+        clearInterval(this.updateInterval);
+        clearTimeout(this.forceUpdateTimeout);
 
-        if (fs.existsSync(tokenFile)) {
-            try {
-                this.log.debug(`Daikin Cloud tokenset found at ${tokenFile}`);
-                tokenSet = JSON.parse(fs.readFileSync(tokenFile).toString());
-            } catch (e) {
-                this.log.debug(`Daikin Cloud could not get tokenset: ${e}`);
-            }
-        }
-
-        const daikinCloud: DaikinCloud = new DaikinCloudController(tokenSet, options);
-
-        daikinCloud.on('token_update', tokenSet => {
-            this.log.info('Retrieved new credentials from Daikin Cloud');
-            fs.writeFileSync(tokenFile, JSON.stringify(tokenSet));
-        });
-
-        try {
-            await daikinCloud.login(username, password);
-        } catch (error) {
-            if (error instanceof Error) {
-                error.message = `Failed to login to Daikin Cloud with ${username}: ${error.message}\nIf the error is about "Missing required fields for registration" try to logout and log back into the Onecta app. Doing this should require you to fill in the missing fields.`;
-            }
-            throw error;
-        }
-
-        return daikinCloud;
+        this.forceUpdateTimeout = setTimeout(async () => {
+            await this.updateDevices();
+            this.startUpdateDevicesInterval();
+        }, delay);
     }
 
-    private isExcludedDevice(excludedDevicesByDeviceId: Array<string>, deviceId): boolean {
+    private startUpdateDevicesInterval() {
+        this.log.debug(`Starting update devices interval every ${this.updateIntervalDelay / ONE_MINUTE} minutes`);
+        this.updateInterval = setInterval(async () => {
+            await this.updateDevices();
+        }, this.updateIntervalDelay);
+    }
+
+    private isExcludedDevice(excludedDevicesByDeviceId: Array<string>, deviceId: string): boolean {
         return typeof excludedDevicesByDeviceId !== 'undefined' && excludedDevicesByDeviceId.includes(deviceId);
+    }
+
+    private getPrivacyFriendlyConfig(config: PlatformConfig): object {
+        return {
+            ...config,
+            clientId: StringUtils.mask(config.clientId),
+            clientSecret: StringUtils.mask(config.clientSecret),
+            excludedDevicesByDeviceId: config.excludedDevicesByDeviceId.map(deviceId => StringUtils.mask(deviceId)),
+        };
     }
 }