+++ changelog.md
@@ -14,6 +14,43 @@ TBD
 
 ### Fixed
 
+## [2.3.0] - 2024-07-02  - 2024-07-04
+
+### Added
+
+- Update daikin-cloud-controller: the controller will now block communication to Daikin when it detects that you have hit your rate limit. If this is not
+  done you will start consuming calls from your next rate limit window. See: https://github.com/Apollon77/daikin-controller-cloud/pull/147
+
+## [2.2.0] - 2024-07-02  - 2024-07-04
+
+### Added
+
+### Changed
+
+- Changed the way we poll the Daikin API, see readme for more details on the current logic
+
+### Fixed
+
+## [2.0.0 - 2.1.0] - 2024-07-02  - 2024-07-03
+
+### Added
+
+- Switch to using the new Daikin Cloud API. (!) PLEASE READ THE README AND CHANGE YOUR CONFIG (!)
+
+### Changed
+
+### Fixed
+
+## [1.8.0-beta.0] - 2023-09-05
+
+### Added
+
+- Support for Altherma (fixes https://github.com/JeroenVdb/homebridge-daikin-cloud/issues/30)
+
+### Changed
+
+### Fixed
+
 ## [1.7.3] - 2024-03-10
 
 ### Added